requirements.txt
ace-tools==0.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
dataclasses-json==0.6.7
duckdb==1.3.1
faiss-cpu==1.11.0
filelock==3.18.0
frozenlist==1.6.2
fsspec==2025.5.1
greenlet==3.2.3
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.4
idna==3.10
jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-0.3.26
langchain-community-0.3.26
langchain-core==0.3.63
langchain-huggingface==0.2.0
langchain-ollama==0.3.3
langchain-text-splitters==0.3.8
langsmith==0.3.45
markupsafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.4.4
mypy-extensions==1.1.0
networkx==3.5
numpy==2.2.6
ollama==0.5.1
orjson==3.10.18
packaging==24.2
pandas==2.3.0
pillow==11.2.1
propcache==0.3.1
pydantic==2.11.5
pydantic-core==2.33.2
pydantic-settings==2.9.1
python-dotenv==1.1.0
pytz==2025.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
langchain-huggingface==0.2.0
sniffio==1.3.1
sqlalchemy==2.0.41
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.4.0
yarl==1.20.0
zstandard==0.23.0